midway logo spinning, needs cleanup

+++ src/game/map.h
@@ -7,9 +7,9 @@
 
 // MAP.ASM
 // asm: 	.bss	MAPPAL13,1	;FL multiplier
-extern int MAPPAL13;
+extern float MAPPAL13;
 // asm: 	.bss	MAPPAL24,1	;FL multiplier
-extern int MAPPAL24;
+extern float MAPPAL24;
 // asm: 	.bss	STOPWATCH,1		;the counter
 extern int STOPWATCH;
 // asm: 	.bss	STOPWATCH_CNTL,1	;the controller	0 = stop, 1 = go
